Alot;
Made Calibration.java. This class will calibrate the boundaries of a color given an good (close-range) picture and a bad (high-range) picture. These boundaries can be used by the TargetFinder since they are made static. 
Made ImageFilter.java. This class will filter and image and represent an histogram of the RGB-values in the picture. This class is also capable of setting boundaries.

Added a FilterTest. You can run it but you will probably get no output, since you don't have the pictures needed. Feel free to test it with your own pictures though.